fixes: warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/branches/bug_fix@82586 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsdata/src/AimsTex2Vol/main.cc
@@ -176,10 +176,10 @@ bool CxParcel::parcellation( AimsData<T> & gmdata )
     }
     if(pmin != numeric_limits<unsigned>::max())
     {
-      x=round(gm[g][0]/sx);
-      y=round(gm[g][1]/sy);
-      z=round(gm[g][2]/sz);
-      parcel(x,y,z,0) = (round) (itex.item(pmin));
+      x=(int)round(gm[g][0]/sx);
+      y=(int)round(gm[g][1]/sy);
+      z=(int)round(gm[g][2]/sz);
+      parcel(x,y,z,0) = (short int)(round) (itex.item(pmin));
     }
     else
       printf("ERROR: could not find any single vertex closer than 1m to (%.0f, %.0f, %.f)!\n",
